# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MOSAP3 Procurement API",
        "description": "Documentação oficial da API do Sistema de Procurement MOSAP3.\n * \n *      Roles:\n *      - **Admin**: Acesso total ao sistema.\n *      - **Procurement Technician**: Gestão de cotações e fornecedores.\n *      - **Public (Fornecedor)**: Acesso via token único enviado por email.",
        "contact": {
            "email": "admin@mosap3.ao"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/api/supplier/register/{token}": {
            "post": {
                "tags": [
                    "Registo Público (Email)"
                ],
                "summary": "Completar Registo de Fornecedor",
                "description": "Permite ao fornecedor preencher os dados e concluir o registo iniciado pelo convite.",
                "operationId": "25b5c8af93911eabd605638622090e83",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "legal_name",
                                    "commercial_name",
                                    "phone",
                                    "nif",
                                    "activity_type",
                                    "province",
                                    "municipality",
                                    "commercial_certificate",
                                    "nif_proof",
                                    "categories"
                                ],
                                "properties": {
                                    "legal_name": {
                                        "type": "string"
                                    },
                                    "commercial_name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "nif": {
                                        "type": "string"
                                    },
                                    "activity_type": {
                                        "type": "string"
                                    },
                                    "province": {
                                        "type": "string"
                                    },
                                    "municipality": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "commercial_certificate": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "commercial_license": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "nif_proof": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "pacto_social": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "non_debtor_certificate": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "product_list": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registo concluído com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Token inválido"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}